# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Seydlitz Dataslate API
  description: |
    API сервера: синхронизация персонажей и раздача пакетов контента.
    Форматы данных — JSON Schema в этой же папке; здесь на них ссылаются через $ref.

    Авторизация — токен сессии: веб получает его в HttpOnly-cookie `session`,
    Android — в теле ответа на вход и передаёт в `Authorization: Bearer …`.
    Сессия живёт 30 дней с последнего использования, но не дольше 90 дней со входа.
    Ответ 401 содержит `WWW-Authenticate: Bearer realm="dataslate"`
    (с `error="invalid_token"`, если токен истёк или недействителен).

    Раз в сутки токен меняется: веб получает новый в cookie, Bearer-клиент — в заголовке
    `X-Session-Token` любого ответа и обязан сохранить его. Старый токен принимается, пока
    новый не использован, и минуту после. Старый токен позже означает кражу: сессия
    отзывается, ответ — `401` с кодом `session_revoked`.

    Запросы с телом — только `Content-Type: application/json`.

    У каждого ответа есть заголовок `X-Request-ID` — по нему в логах сервера находятся
    все записи этого запроса; клиентам стоит показывать его в сообщениях об ошибках.

    Лимиты запросов: 600 в минуту с одного IP (IPv6 — с сети /64) и 300 в минуту
    на пользователя для запросов с сессией. Текущее состояние — в заголовках
    `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset` (unix-время);
    при превышении — `429` с кодом `rate_limited` и `Retry-After`.
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /
    description: Тот же домен, что и веб-сборка (за обратным прокси).
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
  /api/health:
    get:
      operationId: getHealth
      summary: Проверка работоспособности
      tags: [service]
      security: []  # публичный маршрут
      responses:
        "200":
          description: Сервер работает.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
        "4XX":
          $ref: "#/components/responses/Error"

  /api/auth/register:
    post:
      operationId: register
      summary: Регистрация по инвайту
      description: Создаёт пользователя и сразу открывает сессию. Инвайт одноразовый.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite, login, password]
              properties:
                invite:
                  type: string
                  description: Код инвайта; регистр и пробелы по краям не важны.
                login:
                  $ref: "#/components/schemas/Login"
                password:
                  $ref: "#/components/schemas/Password"
                displayName:
                  type: string
                  maxLength: 64
                  description: Отображаемое имя; по умолчанию — логин.
      responses:
        "201":
          $ref: "#/components/responses/Session"
        "400":
          $ref: "#/components/responses/Error"
          description: "`invalid_invite`, `invalid_login`, `invalid_password`, `invalid_display_name`, `bad_request`."
        "409":
          $ref: "#/components/responses/Error"
          description: "`login_taken`. Инвайт при этом не расходуется."
        "429":
          $ref: "#/components/responses/TooManyAttempts"
        "4XX":
          $ref: "#/components/responses/Error"

  /api/auth/login:
    post:
      operationId: login
      summary: Вход
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login:
                  type: string
                  description: Без учёта регистра.
                password:
                  type: string
      responses:
        "200":
          $ref: "#/components/responses/Session"
        "401":
          $ref: "#/components/responses/Error"
          description: "`invalid_credentials` — неверный логин или пароль (что именно — не уточняется)."
        "429":
          $ref: "#/components/responses/TooManyAttempts"
        "4XX":
          $ref: "#/components/responses/Error"

  /api/auth/logout:
    post:
      operationId: logout
      summary: Выход
      description: Закрывает текущую сессию и удаляет cookie.
      tags: [auth]
      responses:
        "204":
          description: Сессия закрыта.
        "4XX":
          $ref: "#/components/responses/Error"

  /api/me:
    get:
      operationId: getMe
      summary: Текущий пользователь
      tags: [auth]
      responses:
        "200":
          description: Пользователь текущей сессии.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "4XX":
          $ref: "#/components/responses/Error"

  /api/me/login:
    put:
      operationId: changeLogin
      summary: Сменить логин
      description: |
        Нужен текущий пароль. Сессии не завершаются. Неверный пароль — `403`, не `401`: сессия действительна.
        Неудачные попытки считаются вместе с попытками входа.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login:
                  $ref: "#/components/schemas/Login"
                password:
                  type: string
                  description: Текущий пароль.
      responses:
        "200":
          description: Логин изменён (или совпадал с текущим).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/Error"
          description: "`invalid_login`."
        "403":
          $ref: "#/components/responses/Error"
          description: "`wrong_password` — неверный текущий пароль."
        "409":
          $ref: "#/components/responses/Error"
          description: "`login_taken`."
        "429":
          $ref: "#/components/responses/TooManyAttempts"
        "4XX":
          $ref: "#/components/responses/Error"

  /api/me/password:
    put:
      operationId: changePassword
      summary: Сменить пароль
      description: |
        Нужен текущий пароль. Все сессии, кроме текущей, завершаются.
        Неверный пароль — `403`, не `401`: сессия действительна. Неудачные попытки считаются вместе с попытками входа.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  $ref: "#/components/schemas/Password"
      responses:
        "204":
          description: Пароль изменён, остальные сессии завершены.
        "400":
          $ref: "#/components/responses/Error"
          description: "`invalid_password` — новый пароль не подходит по длине."
        "403":
          $ref: "#/components/responses/Error"
          description: "`wrong_password` — неверный текущий пароль."
        "429":
          $ref: "#/components/responses/TooManyAttempts"
        "4XX":
          $ref: "#/components/responses/Error"

  /api/invites:
    get:
      operationId: listInvites
      summary: Список инвайтов
      description: Только для администратора. Коды не возвращаются — сервер хранит лишь их хеши.
      tags: [invites]
      responses:
        "200":
          description: Инвайты, новые первыми.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invite"
        "4XX":
          $ref: "#/components/responses/Error"
    post:
      operationId: createInvite
      summary: Создать инвайт
      description: Только для администратора. Код возвращается один раз — в этом ответе.
      tags: [invites]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresInDays:
                  type: integer
                  minimum: 1
                  maximum: 90
                  default: 7
      responses:
        "201":
          description: Инвайт создан.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Invite"
                  - required: [code]
        "400":
          $ref: "#/components/responses/Error"
          description: "`invalid_expiry`."
        "4XX":
          $ref: "#/components/responses/Error"

  /api/sessions:
    get:
      operationId: listSessions
      summary: Активные сессии текущего пользователя
      tags: [auth]
      responses:
        "200":
          description: Сессии, последние использованные первыми.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionInfo"
        "4XX":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteOtherSessions
      summary: Выйти на всех остальных устройствах
      tags: [auth]
      responses:
        "204":
          description: Все сессии, кроме текущей, завершены.
        "4XX":
          $ref: "#/components/responses/Error"

  /api/sessions/{id}:
    delete:
      operationId: deleteSession
      summary: Завершить сессию
      description: Своя сессия; текущая — то же, что выход.
      tags: [auth]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Сессия завершена.
        "404":
          $ref: "#/components/responses/Error"
          description: "`not_found`."
        "4XX":
          $ref: "#/components/responses/Error"

  /api/invites/{id}:
    delete:
      operationId: deleteInvite
      summary: Отозвать инвайт
      description: Только для администратора и только неиспользованный инвайт.
      tags: [invites]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Инвайт удалён.
        "404":
          $ref: "#/components/responses/Error"
          description: "`not_found` — нет такого или уже использован."
        "4XX":
          $ref: "#/components/responses/Error"

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Веб. Cookie ставит сервер при входе и регистрации (HttpOnly, Secure, SameSite=Lax).
    bearerAuth:
      type: http
      scheme: bearer
      description: Android. Токен — поле `token` ответа на вход или регистрацию.

  schemas:
    Health:
      type: object
      required: [status, version]
      properties:
        status:
          const: ok
        version:
          type: string
          description: Версия сборки сервера; "dev" — собран без -ldflags.
          examples: ["0.1.0"]

    Login:
      type: string
      pattern: "^[A-Za-z0-9_.-]{3,32}$"
      description: Уникален без учёта регистра.

    Password:
      type: string
      minLength: 8
      maxLength: 128

    User:
      type: object
      required: [id, login, displayName, isAdmin, createdAt]
      properties:
        id:
          type: string
          format: uuid
        login:
          $ref: "#/components/schemas/Login"
        displayName:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time

    SessionInfo:
      type: object
      required: [id, createdAt, lastUsedAt, userAgent, current]
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
          description: Вход.
        lastUsedAt:
          type: string
          format: date-time
          description: С точностью до часа.
        userAgent:
          type: string
        current:
          type: boolean
          description: Сессия этого запроса.

    Invite:
      type: object
      required: [id, createdAt, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          description: Только в ответе на создание.
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        usedAt:
          type: string
          format: date-time
          description: Нет поля — инвайт не использован.
        usedBy:
          type: string
          description: Логин зарегистрировавшегося (нет поля, если пользователь удалён).

    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Описание ошибки для человека.
          examples: ["Not Found"]
        code:
          type: string
          description: |
            Машиночитаемый код. Общие: `bad_request`, `too_large`, `unsupported_media_type`, `unauthorized`,
            `session_revoked`, `forbidden`, `not_found`, `method_not_allowed`, `rate_limited`,
            `too_many_attempts`, `internal`;
            остальные — в описаниях ответов конкретных операций.
          examples: ["not_found"]

  responses:
    Error:
      description: Ошибка. Для 405 — заголовок Allow.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    TooManyAttempts:
      description: "`too_many_attempts` — слишком много неудачных попыток с этого адреса или для этого логина."
      headers:
        Retry-After:
          description: Через сколько секунд можно повторить.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

    Session:
      description: Сессия открыта; токен также установлен в cookie `session`.
      headers:
        Set-Cookie:
          schema:
            type: string
      content:
        application/json:
          schema:
            type: object
            required: [token, user]
            properties:
              token:
                type: string
              user:
                $ref: "#/components/schemas/User"

tags:
  - name: service
    description: Служебные маршруты.
  - name: auth
    description: Регистрация, вход, сессии.
  - name: invites
    description: Инвайты на регистрацию (только администратор).